backend/requirements.txt
fastapi[standard]==0.115.*
uvicorn[standard]==0.34.*
sqlalchemy==2.0.*
alembic==1.14.*
psycopg2-binary==2.9.*
python-jose[cryptography]==3.3.*
passlib[bcrypt]==1.7.*
bcrypt==4.0.*
python-multipart==0.0.*
pydantic-settings==2.7.*
httpx==0.28.*
firebase-admin==7.2.*
croniter==5.0.*
